devices: Add memctl device
Memctl provides a way for the guest to control its physical memory
properties, and enables optimizations and security features. For
example, the guest can provide information to the host where parts of a
hugepage may be unbacked, or sensitive data may not be swapped out, etc.

Memctl allows guests to manipulate its gPTE entries in the SLAT, and
also some other properties of the memory map the back's host memory.
This is achieved by using the KVM_CAP_SYNC_MMU capability. When this
capability is available, the changes in the backing of the memory region
on the host are automatically reflected into the guest. For example, an
mmap() or madvise() that affects the region will be made visible
immediately.

There are two components of the implementation: the guest Linux driver
and Virtual Machine Monitor (VMM) backend. They communicate with each
other via a virtio virtqueue.

The Cloud Hypervisor component can be enabled with --memctl.

A proof of concept program using the userspace interface to set
a hugetlb anon vma name can be found at:
https://github.com/Dummyc0m/memctl-set-anon-vma-name

Signed-off-by: Yuanchu Xie <yuanchu@google.com>